refactor tree click handler
